--- judging-algorithm/.gitignore
@@ -1,2 +1,0 @@
-# dependencies
-/node_modules
