--- judging-algorithm/data/csv_inputs/hackers.csv
@@ -1,36 +1,0 @@
-name,liveDemo
-Feed the Desert,yes
-Pyramid Scheme,yes
-SunShare,yes
-BASS Booster,yes
-Boston 101,yes
-Recipe Generator,yes
-Boston Housing Analysis,no
-BPWorks - Boston Public Works,yes
-Outreach,yes
-Hello World,yes
-Northeastern CEO,yes
-Overgrown - Prune Your Tabs!,yes
-KdramaUniverse,no
-Explore,yes
-Python Search Algorithm Visualization,yes
-ThriftBuddy,no
-WiseWallet,yes
-Roommate Hub,yes
-Askle,yes
-Connect 4 Companion,yes
-personal decision making assistance,yes
-FormFusion,yes
-Astro,yes
-Personify,yes
-FreshFarm,yes
-Pick-Your-Path,no
-Cactus Study,yes
-SNAPpy,yes
-Is that a challenge,yes
-Time And Place (TAP),yes
-tl;dr,no
-EverythingHoops,yes
-Carson,yes
-Evnts,yes
-Page 14,yes
