--- judging-algorithm/data/csv_inputs/judges.csv
@@ -1,8 +1,0 @@
-name
-Allison Alder
-Devan Gross
-William Guo
-Cari Liu
-Nupur Neogi
-Anika Rabenhorst
-Alex Takayama
