--- judging-algorithm/data/csv_inputs/rooms.csv
@@ -1,4 +1,0 @@
-name
-core canyon
-dust bowl
-nomad nook
