--- judging-algorithm/data/csv_templates/hackers_example.csv
@@ -1,36 +1,0 @@
-name,liveDemo
-team1,TRUE
-team2,FALSE
-team3,TRUE
-team4,TRUE
-team5,TRUE
-team6,FALSE
-team7,FALSE
-team8,TRUE
-team9,TRUE
-team10,FALSE
-team11,FALSE
-team12,TRUE
-team13,TRUE
-team14,TRUE
-team15,TRUE
-team16,TRUE
-team17,TRUE
-team18,TRUE
-team19,TRUE
-team20,TRUE
-team21,TRUE
-team22,TRUE
-team23,FALSE
-team24,TRUE
-team25,TRUE
-team26,TRUE
-team27,TRUE
-team28,TRUE
-team29,TRUE
-team30,TRUE
-team31,TRUE
-team32,TRUE
-team33,FALSE
-team34,FALSE
-team35,TRUE
