--- judging-algorithm/data/csv_templates/judges_example.csv
@@ -1,5 +1,0 @@
-name
-lisa
-surbhi
-judy
-david
