--- judging-algorithm/data/csv_templates/judges_template.csv
@@ -1,1 +1,0 @@
-name
