--- judging-algorithm/data/csv_templates/rooms_example.csv
@@ -1,4 +1,0 @@
-name
-hidden figures
-board room
-rob’s office
