--- judging-algorithm/data/csv_templates/rooms_template.csv
@@ -1,1 +1,0 @@
-name
