--- judging-algorithm/data/json_outputs/hackerResults.json
@@ -1,1 +1,0 @@
-[{"project":"Feed the Desert","time":"10:00","judges":["Allison Alder","Devan Gross","Alex Takayama"],"room":"core canyon"},{"project":"Pyramid Scheme","time":"10:00","judges":["William Guo","Cari Liu"],"room":"dust bowl"},{"project":"SunShare","time":"10:00","judges":["Nupur Neogi","Anika Rabenhorst"],"room":"nomad nook"},{"project":"BASS Booster","time":"10:09","judges":["Allison Alder","Devan Gross","Alex Takayama"],"room":"core canyon"},{"project":"Boston 101","time":"10:09","judges":["William Guo","Cari Liu"],"room":"dust bowl"},{"project":"Recipe Generator","time":"10:09","judges":["Nupur Neogi","Anika Rabenhorst"],"room":"nomad nook"},{"project":"BPWorks - Boston Public Works","time":"10:18","judges":["Allison Alder","Devan Gross","Alex Takayama"],"room":"core canyon"},{"project":"Outreach","time":"10:18","judges":["William Guo","Cari Liu"],"room":"dust bowl"},{"project":"Hello World","time":"10:18","judges":["Nupur Neogi","Anika Rabenhorst"],"room":"nomad nook"},{"project":"Northeastern CEO","time":"10:27","judges":["Allison Alder","Devan Gross","Alex Takayama"],"room":"core canyon"},{"project":"Overgrown - Prune Your Tabs!","time":"10:27","judges":["William Guo","Cari Liu"],"room":"dust bowl"},{"project":"Explore","time":"10:27","judges":["Nupur Neogi","Anika Rabenhorst"],"room":"nomad nook"},{"project":"Python Search Algorithm Visualization","time":"10:36","judges":["Allison Alder","Devan Gross","Alex Takayama"],"room":"core canyon"},{"project":"WiseWallet","time":"10:36","judges":["William Guo","Cari Liu"],"room":"dust bowl"},{"project":"Roommate Hub","time":"10:36","judges":["Nupur Neogi","Anika Rabenhorst"],"room":"nomad nook"},{"project":"Askle","time":"10:45","judges":["Allison Alder","Devan Gross","Alex Takayama"],"room":"core canyon"},{"project":"Connect 4 Companion","time":"10:45","judges":["William Guo","Cari Liu"],"room":"dust bowl"},{"project":"personal decision making assistance","time":"10:45","judges":["Nupur Neogi","Anika Rabenhorst"],"room":"nomad nook"},{"project":"FormFusion","time":"10:54","judges":["Allison Alder","Devan Gross","Alex Takayama"],"room":"core canyon"},{"project":"Astro","time":"10:54","judges":["William Guo","Cari Liu"],"room":"dust bowl"},{"project":"Personify","time":"10:54","judges":["Nupur Neogi","Anika Rabenhorst"],"room":"nomad nook"},{"project":"FreshFarm","time":"11:03","judges":["Allison Alder","Devan Gross","Alex Takayama"],"room":"core canyon"},{"project":"Cactus Study","time":"11:03","judges":["William Guo","Cari Liu"],"room":"dust bowl"},{"project":"SNAPpy","time":"11:03","judges":["Nupur Neogi","Anika Rabenhorst"],"room":"nomad nook"},{"project":"Is that a challenge","time":"11:12","judges":["Allison Alder","Devan Gross","Alex Takayama"],"room":"core canyon"},{"project":"Time And Place (TAP)","time":"11:12","judges":["William Guo","Cari Liu"],"room":"dust bowl"},{"project":"EverythingHoops","time":"11:12","judges":["Nupur Neogi","Anika Rabenhorst"],"room":"nomad nook"},{"project":"Carson","time":"11:21","judges":["Allison Alder","Devan Gross","Alex Takayama"],"room":"core canyon"},{"project":"Evnts","time":"11:21","judges":["William Guo","Cari Liu"],"room":"dust bowl"},{"project":"Page 14","time":"11:21","judges":["Nupur Neogi","Anika Rabenhorst"],"room":"nomad nook"}]
