--- judging-algorithm/data/json_outputs/judgeResults.json
@@ -1,1 +1,0 @@
-[{"room":"core canyon","judge":"Allison Alder","projects":[{"room":"core canyon","judge":"Allison Alder","time":"10:00","project":"Feed the Desert"},{"room":"core canyon","judge":"Allison Alder","time":"10:09","project":"BASS Booster"},{"room":"core canyon","judge":"Allison Alder","time":"10:18","project":"BPWorks - Boston Public Works"},{"room":"core canyon","judge":"Allison Alder","time":"10:27","project":"Northeastern CEO"},{"room":"core canyon","judge":"Allison Alder","time":"10:36","project":"Python Search Algorithm Visualization"},{"room":"core canyon","judge":"Allison Alder","time":"10:45","project":"Askle"},{"room":"core canyon","judge":"Allison Alder","time":"10:54","project":"FormFusion"},{"room":"core canyon","judge":"Allison Alder","time":"11:03","project":"FreshFarm"},{"room":"core canyon","judge":"Allison Alder","time":"11:12","project":"Is that a challenge"},{"room":"core canyon","judge":"Allison Alder","time":"11:21","project":"Carson"}]},{"room":"core canyon","judge":"Devan Gross","projects":[{"room":"core canyon","judge":"Devan Gross","time":"10:00","project":"Feed the Desert"},{"room":"core canyon","judge":"Devan Gross","time":"10:09","project":"BASS Booster"},{"room":"core canyon","judge":"Devan Gross","time":"10:18","project":"BPWorks - Boston Public Works"},{"room":"core canyon","judge":"Devan Gross","time":"10:27","project":"Northeastern CEO"},{"room":"core canyon","judge":"Devan Gross","time":"10:36","project":"Python Search Algorithm Visualization"},{"room":"core canyon","judge":"Devan Gross","time":"10:45","project":"Askle"},{"room":"core canyon","judge":"Devan Gross","time":"10:54","project":"FormFusion"},{"room":"core canyon","judge":"Devan Gross","time":"11:03","project":"FreshFarm"},{"room":"core canyon","judge":"Devan Gross","time":"11:12","project":"Is that a challenge"},{"room":"core canyon","judge":"Devan Gross","time":"11:21","project":"Carson"}]},{"room":"dust bowl","judge":"William Guo","projects":[{"room":"dust bowl","judge":"William Guo","time":"10:00","project":"Pyramid Scheme"},{"room":"dust bowl","judge":"William Guo","time":"10:09","project":"Boston 101"},{"room":"dust bowl","judge":"William Guo","time":"10:18","project":"Outreach"},{"room":"dust bowl","judge":"William Guo","time":"10:27","project":"Overgrown - Prune Your Tabs!"},{"room":"dust bowl","judge":"William Guo","time":"10:36","project":"WiseWallet"},{"room":"dust bowl","judge":"William Guo","time":"10:45","project":"Connect 4 Companion"},{"room":"dust bowl","judge":"William Guo","time":"10:54","project":"Astro"},{"room":"dust bowl","judge":"William Guo","time":"11:03","project":"Cactus Study"},{"room":"dust bowl","judge":"William Guo","time":"11:12","project":"Time And Place (TAP)"},{"room":"dust bowl","judge":"William Guo","time":"11:21","project":"Evnts"}]},{"room":"dust bowl","judge":"Cari Liu","projects":[{"room":"dust bowl","judge":"Cari Liu","time":"10:00","project":"Pyramid Scheme"},{"room":"dust bowl","judge":"Cari Liu","time":"10:09","project":"Boston 101"},{"room":"dust bowl","judge":"Cari Liu","time":"10:18","project":"Outreach"},{"room":"dust bowl","judge":"Cari Liu","time":"10:27","project":"Overgrown - Prune Your Tabs!"},{"room":"dust bowl","judge":"Cari Liu","time":"10:36","project":"WiseWallet"},{"room":"dust bowl","judge":"Cari Liu","time":"10:45","project":"Connect 4 Companion"},{"room":"dust bowl","judge":"Cari Liu","time":"10:54","project":"Astro"},{"room":"dust bowl","judge":"Cari Liu","time":"11:03","project":"Cactus Study"},{"room":"dust bowl","judge":"Cari Liu","time":"11:12","project":"Time And Place (TAP)"},{"room":"dust bowl","judge":"Cari Liu","time":"11:21","project":"Evnts"}]},{"room":"nomad nook","judge":"Nupur Neogi","projects":[{"room":"nomad nook","judge":"Nupur Neogi","time":"10:00","project":"SunShare"},{"room":"nomad nook","judge":"Nupur Neogi","time":"10:09","project":"Recipe Generator"},{"room":"nomad nook","judge":"Nupur Neogi","time":"10:18","project":"Hello World"},{"room":"nomad nook","judge":"Nupur Neogi","time":"10:27","project":"Explore"},{"room":"nomad nook","judge":"Nupur Neogi","time":"10:36","project":"Roommate Hub"},{"room":"nomad nook","judge":"Nupur Neogi","time":"10:45","project":"personal decision making assistance"},{"room":"nomad nook","judge":"Nupur Neogi","time":"10:54","project":"Personify"},{"room":"nomad nook","judge":"Nupur Neogi","time":"11:03","project":"SNAPpy"},{"room":"nomad nook","judge":"Nupur Neogi","time":"11:12","project":"EverythingHoops"},{"room":"nomad nook","judge":"Nupur Neogi","time":"11:21","project":"Page 14"}]},{"room":"nomad nook","judge":"Anika Rabenhorst","projects":[{"room":"nomad nook","judge":"Anika Rabenhorst","time":"10:00","project":"SunShare"},{"room":"nomad nook","judge":"Anika Rabenhorst","time":"10:09","project":"Recipe Generator"},{"room":"nomad nook","judge":"Anika Rabenhorst","time":"10:18","project":"Hello World"},{"room":"nomad nook","judge":"Anika Rabenhorst","time":"10:27","project":"Explore"},{"room":"nomad nook","judge":"Anika Rabenhorst","time":"10:36","project":"Roommate Hub"},{"room":"nomad nook","judge":"Anika Rabenhorst","time":"10:45","project":"personal decision making assistance"},{"room":"nomad nook","judge":"Anika Rabenhorst","time":"10:54","project":"Personify"},{"room":"nomad nook","judge":"Anika Rabenhorst","time":"11:03","project":"SNAPpy"},{"room":"nomad nook","judge":"Anika Rabenhorst","time":"11:12","project":"EverythingHoops"},{"room":"nomad nook","judge":"Anika Rabenhorst","time":"11:21","project":"Page 14"}]},{"room":"core canyon","judge":"Alex Takayama","projects":[{"room":"core canyon","judge":"Alex Takayama","time":"10:00","project":"Feed the Desert"},{"room":"core canyon","judge":"Alex Takayama","time":"10:09","project":"BASS Booster"},{"room":"core canyon","judge":"Alex Takayama","time":"10:18","project":"BPWorks - Boston Public Works"},{"room":"core canyon","judge":"Alex Takayama","time":"10:27","project":"Northeastern CEO"},{"room":"core canyon","judge":"Alex Takayama","time":"10:36","project":"Python Search Algorithm Visualization"},{"room":"core canyon","judge":"Alex Takayama","time":"10:45","project":"Askle"},{"room":"core canyon","judge":"Alex Takayama","time":"10:54","project":"FormFusion"},{"room":"core canyon","judge":"Alex Takayama","time":"11:03","project":"FreshFarm"},{"room":"core canyon","judge":"Alex Takayama","time":"11:12","project":"Is that a challenge"},{"room":"core canyon","judge":"Alex Takayama","time":"11:21","project":"Carson"}]}]
