--- judging-algorithm/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "judging-algorithm",
-  "version": "1.0.0",
-  "description": "Judging schedule for HackBeanpot",
-  "main": "index.ts",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "surbhi gulati, lisa jiang",
-  "license": "ISC",
-  "dependencies": {
-    "@types/node": "^18.11.18",
-    "csv": "^6.3.0",
-    "csv-parse": "^5.3.10",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.0.4",
-    "write-json-file": "^5.0.0"
-  },
-  "type": "module"
-}
