--- judging-algorithm/templates/hackers_example.csv
@@ -1,36 +1,0 @@
-name,liveDemo,devpostLink
-team1,TRUE,www.google.com/cats
-team2,FALSE,www.google.com/dogs
-team3,TRUE,www.google.com/apple
-team4,TRUE,www.google.com/mango
-team5,TRUE,www.google.com/kangaroo
-team6,FALSE,www.google.com/elephant
-team7,FALSE,www.google.com/table
-team8,TRUE,www.google.com/chair
-team9,TRUE,www.google.com/phone
-team10,FALSE,www.google.com/laptop
-team11,FALSE,www.google.com/shirt
-team12,TRUE,www.google.com/sweater
-team13,TRUE,www.google.com/pants
-team14,TRUE,www.google.com/jeans
-team15,TRUE,www.google.com/coat
-team16,TRUE,www.google.com/blazer
-team17,TRUE,www.google.com/beanie
-team18,TRUE,www.google.com/hanger
-team19,TRUE,www.google.com/ring
-team20,TRUE,www.google.com/bracelet
-team21,TRUE,www.google.com/necklace
-team22,TRUE,www.google.com/earring
-team23,FALSE,www.google.com/hackbeanpot
-team24,TRUE,www.google.com/running
-team25,TRUE,www.google.com/out
-team26,TRUE,www.google.com/of
-team27,TRUE,www.google.com/random
-team28,TRUE,www.google.com/word
-team29,TRUE,www.google.com/ideas
-team30,TRUE,www.google.com/for
-team31,TRUE,www.google.com/this
-team32,TRUE,www.google.com/testfile
-team33,FALSE,www.google.com/techteamrocks
-team34,FALSE,www.google.com/letsdothis
-team35,TRUE,www.google.com/okpleaseworkout
