--- judging-algorithm/templates/hackers_template.csv
@@ -1,1 +1,0 @@
-name,liveDemo,devpostLink
