--- judging-algorithm/templates/judges_example.csv
@@ -1,5 +1,0 @@
-name,inPerson,company
-lisa,TRUE,meta
-surbhi,TRUE,google
-judy,TRUE,toast
-david,FALSE,amazon
