--- judging-algorithm/templates/judges_template.csv
@@ -1,1 +1,0 @@
-name,inPerson,company
