--- judging-algorithm/templates/rooms_example.csv
@@ -1,4 +1,0 @@
-name,capacity
-hidden figures,3
-board room,3
-rob’s office,1
