--- judging-algorithm/tsconfig.json
@@ -1,43 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "esnext",
-    "lib": ["dom", "esnext"],
-    "jsx": "react",
-    "module": "ESNext",
-    "moduleResolution": "Node",
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "strict": true,
-    "skipLibCheck": true,
-    "noImplicitAny": true,
-    "noFallthroughCasesInSwitch": true,
-    "noImplicitOverride": true,
-    "noImplicitReturns": true,
-    "noImplicitThis": true,
-    "noUnusedParameters": true,
-    "allowUnreachableCode": true,
-    "allowUnusedLabels": true,
-    "exactOptionalPropertyTypes": false,
-    "strictFunctionTypes": true,
-    "strictNullChecks": true,
-    "strictPropertyInitialization": true, 
-    "resolveJsonModule": true,
-  },
-  "include": [
-    "main-site", 
-    "live-site", 
-    "shared-ui", 
-    "judging-algorithm", 
-    "*.json"
-  ],
-  "exclude": [
-    "node_modules",
-    "public",
-    "gatsby-node.ts",
-    "gatsby-config.ts",
-    "plugins/**/*"
-  ],
-  "ts-node": {
-    "esm": true
-  }
-}
